fix: offer-page ownership check + countdown on SQL Server
Caught via browser verification on staging:
- /offers/{token} 403'd the legitimate provider because pdo_sqlsrv (Railway)
  returns bigint columns as strings, so the strict $provider->id !==
  $offer->provider_id check failed (local dblib returns ints, so tests passed).
  Cast both sides to int.
- My Offers countdown showed fractional seconds (Carbon 3 diffInSeconds returns
  a float); cast to int.

+++ resources/views/filament/dashboard/pages/my-offers.blade.php
@@ -18,7 +18,7 @@
             @foreach ($pending as $offer)
                 @php
                     $intake = $offer->intakeRequest;
-                    $secondsLeft = ($offer->expires_at && $offer->expires_at->isFuture()) ? now()->diffInSeconds($offer->expires_at) : 0;
+                    $secondsLeft = ($offer->expires_at && $offer->expires_at->isFuture()) ? (int) now()->diffInSeconds($offer->expires_at) : 0;
                 @endphp
                 <div class="{{ $card }}" wire:key="offer-{{ $offer->id }}">
                     <div class="flex flex-wrap items-start justify-between gap-4">
